requirement.txt
Flask==3.1.0

python-dotenv==1.0.1

gunicorn==23.0.0

twilio==9.3.6